ci: enable ccache for MSVC and Meson, fix MSVC conversion warnings
- Set CMAKE_MSVC_DEBUG_INFORMATION_FORMAT=Embedded (/Z7) when ccache is enabled on MSVC so compilation outputs are cacheable.
- Enable -Db_ccache=true in Meson configure step.
- Fix MSVC truncation/conversion warnings in examples and gridbox_test.

+++ examples/component/homescreen.cpp
@@ -415,7 +415,7 @@ int main() {
   // ---------------------------------------------------------------------------
   // Paragraph
   // ---------------------------------------------------------------------------
-  auto make_box = [](size_t dimx, size_t dimy) {
+  auto make_box = [](int dimx, int dimy) {
     std::string title = std::to_string(dimx) + "x" + std::to_string(dimy);
     return window(text(title) | hcenter | bold,
                   text("content") | hcenter | dim) |
